renamed internal/pkg/caaspctl/deployments/ssh/cilium.go -> internal/pkg/caaspctl/cni/cilium.go
